Fixing type for SW c file.

+++ device/perform_ls.c
@@ -7,7 +7,7 @@
 // --------------------------------------------------------------------------
 // --------------------------------------------------------------------------
 void perform_ls(
-			uint				DockConst_max_num_of_iters,
+			ushort				DockConst_max_num_of_iters,
 			float				DockConst_rho_lower_bound,
 			float				DockConst_base_dmov_mul_sqrt3,
 			uchar				DockConst_num_of_genes,
